fixes for built version + e2e test
