+++ camagru/_tests_/component/SigninForm.test.tsx
@@ -1,14 +1,13 @@
-import { render, screen } from '@testing-library/react'
+import { render, screen, waitFor } from '@testing-library/react'
 import SignInForm from '@/components/SigninForm'
 import userEvent from '@testing-library/user-event'
-// import { signIn, getSession } from 'next-auth/react'
+import { signIn } from 'next-auth/react'
 
-jest.mock('next-auth/react')
+jest.mock('next-auth/react', () => ({
+  signIn: jest.fn(),
+}))
 
-// test with loginin and redirection to /user/userid
-// if login ing fails redirection to /
 // fail login
-
 // check that nav bar is not on the screen
 
 describe('SigninForm', () => {
@@ -32,7 +31,7 @@ describe('SigninForm', () => {
     expect(submitButton).toBeDisabled()
   })
 
-  it('subnit button is active when fields are fielled', async () => {
+  it('submit button is active when fields are fielled', async () => {
     render(<SignInForm />)
 
     const loginInput = screen.getByTestId('login-signin')
@@ -43,5 +42,43 @@ describe('SigninForm', () => {
     await userEvent.type(passwordInput, 'StrongPass!1234')
 
     expect(submitButton).not.toBeDisabled()
+  })
+
+  it('login and password are empty', async () => {
+    render(<SignInForm />)
+
+    const submitButton = screen.getByTestId('signin-button')
+
+    expect(submitButton).toBeDisabled()
+    submitButton.removeAttribute('disabled')
+    expect(submitButton).not.toBeDisabled()
+
+    await userEvent.click(submitButton)
+
+    expect(
+      await screen.findByText('Please enter your email or username')
+    ).toBeInTheDocument()
+    expect(
+      await screen.findByText('Please enter your password')
+    ).toBeInTheDocument()
+  })
+
+  it('show error when credentals are invalid', async () => {
+    ;(signIn as jest.Mock).mockResolvedValueOnce({
+      error: 'Credentialssignin',
+    })
+    render(<SignInForm />)
+    const loginInput = screen.getByTestId('login-signin')
+    const passwordInput = screen.getByTestId('password-signin')
+    const submitButton = screen.getByTestId('signin-button')
+
+    await userEvent.type(loginInput, 'wronguser')
+    await userEvent.type(passwordInput, 'WrongPass123!')
+
+    await userEvent.click(submitButton)
+
+    await waitFor(() => {
+      expect(screen.getByText(/Credentialssignin/i)).toBeInTheDocument()
+    })
   })
 })
