+++ camagru/package.json
@@ -3,9 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
+    "build:test": "cross-env NODE_ENV=test next build",
+    "dev:test": "cross-env NODE_ENV=test next dev -p 3001",
     "dev": "next dev --turbopack",
     "test:components": "jest",
-    "test:e2e":"npx playwright test --ui",
+    "test:e2e": "npx playwright test --ui",
     "coverage": "jest --coverage",
     "build": "next build",
     "start": "next start",
@@ -40,6 +42,7 @@
     "@types/react-dom": "^19.1.6",
     "@types/uuid": "^10.0.0",
     "@types/zxcvbn": "^4.4.5",
+    "cross-env": "^7.0.3",
     "eslint": "^9",
     "eslint-config-next": "15.3.3",
     "jest": "^30.0.0",
