+++ camagru/styles/Profile.module.css
@@ -92,7 +92,7 @@
   cursor: pointer;
 }
 
-button:disabled {
+.button:disabled {
   opacity: 0.5;
   cursor: not-allowed;
 }
