+++ camagru/styles/Register.module.css
@@ -62,7 +62,7 @@
   cursor: pointer;
 }
 
-button:disabled {
+.button:disabled {
   opacity: 0.5;
   cursor: not-allowed;
 }
