+++ camagru/styles/Settings.module.css
@@ -61,7 +61,7 @@
   cursor: pointer;
 }
 
-button:disabled {
+.button:disabled {
   opacity: 0.5;
   cursor: not-allowed;
 }
